ai-agent-python/pyproject.toml
[project]
name = "ai-agent-python"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "langchain==1.2.0",
    "langgraph==1.0.5",
    "langsmith>=0.4.49",
    "openai>=1.68.2,<2.0.0",
    "fastapi>=0.115.5,<1.0.0",
    "uvicorn>=0.29.0,<1.0.0",
    "python-dotenv>=1.0.0,<2.0.0",
    "langgraph-cli[inmem]>=0.4.11",
    "langchain-openai>=1.1.0",
    "copilotkit>=0.1.77",
    "langgraph-api>=0.7.16",
    "langchain-mcp-adapters>=0.2.1",
    "tavily-python>=0.7.22",
    "langchain-community>=0.4.1",
    "json5>=0.13.0",
    "langchain-core>=1.2.12",
    "deepagents>=0.3.3",
    "pyjwt>=2.11.0",
]

